Update index.html
Let's stick to English for code, comments and variables please. ;-)
I changed the description of the collective to something that is closer to what we decided when we started the project. What we are, what we want and how we introduce ourselves to the world should only be changed through a collective discussion on both mailing lists!
Please limit the number of columns to something around 80 characters, very long lines of code are difficult to read.